+++ cmd/chainctl/app/config_flag.go
@@ -1,0 +1,14 @@
+package app
+
+import (
+	"github.com/spf13/cobra"
+
+	"github.com/dobrovols/chainctl/cmd/chainctl/declarative"
+)
+
+func markDeclarative(cmd *cobra.Command) {
+	if cmd.Annotations == nil {
+		cmd.Annotations = map[string]string{}
+	}
+	cmd.Annotations[declarative.AnnotationEnabled] = "true"
+}